create mock for localStorage and split update tests
